// src/queries/gen-types.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** An ISO-8601 encoded date string. */
  Date: any;
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any;
  /** A Git object ID. */
  GitObjectID: any;
  /** Git SSH string */
  GitSSHRemote: any;
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any;
  /** A string containing HTML code. */
  HTML: any;
  /** An ISO-8601 encoded UTC date string with millisecond precision. */
  PreciseDateTime: any;
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any;
  /** A valid x509 certificate string */
  X509Certificate: any;
};
// ...
export type FollowerConnection = {
  /** A list of edges. */
  edges: Maybe<Array<Maybe<UserEdge>>>;
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};

/** The connection type for User. */
export type FollowingConnection = {
  /** A list of edges. */
  edges: Maybe<Array<Maybe<UserEdge>>>;
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type Gist = Node & Starrable & UniformResourceLocatable & {
  /** A list of comments associated with the gist */
  comments: GistCommentConnection;
  /** Identifies the date and time when the object was created. */
  createdAt: Scalars['DateTime'];
  /** The gist description. */
  description: Maybe<Scalars['String']>;
  /** The files in this gist. */
  files: Maybe<Array<Maybe<GistFile>>>;
  /** A list of forks associated with the gist */
  forks: GistConnection;
  id: Scalars['ID'];
  /** Identifies if the gist is a fork. */
  isFork: Scalars['Boolean'];
  /** Whether the gist is public or not. */
  isPublic: Scalars['Boolean'];
  /** The gist name. */
  name: Scalars['String'];
  /** The gist owner. */
  owner: Maybe<RepositoryOwner>;
  /** Identifies when the gist was last pushed to. */
  pushedAt: Maybe<Scalars['DateTime']>;
  /** The HTML path to this resource. */
  resourcePath: Scalars['URI'];
  /** Returns a count of how many stargazers there are on this object */
  stargazerCount: Scalars['Int'];
  /** A list of users who have starred this starrable. */
  stargazers: StargazerConnection;
  /** Identifies the date and time when the object was last updated. */
  updatedAt: Scalars['DateTime'];
  /** The HTTP URL for this Gist. */
  url: Scalars['URI'];
  /** Returns a boolean indicating whether the viewing user has starred this starrable. */
  viewerHasStarred: Scalars['Boolean'];
};
// ...
export type GistCommentConnection = {
  /** A list of edges. */
  edges: Maybe<Array<Maybe<GistCommentEdge>>>;
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<GistComment>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};

/** An edge in a connection. */
export type GistCommentEdge = {
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
  /** The item at the end of the edge. */
  node: Maybe<GistComment>;
};

/** The connection type for Gist. */
export type GistConnection = {
  /** A list of edges. */
  edges: Maybe<Array<Maybe<GistEdge>>>;
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<Gist>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};

/** An edge in a connection. */
export type GistEdge = {
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
  /** The item at the end of the edge. */
  node: Maybe<Gist>;
};

/** A file in a gist. */
export type GistFile = {
  /** The file name encoded to remove characters that are invalid in URL paths. */
  encodedName: Maybe<Scalars['String']>;
  /** The gist file encoding. */
  encoding: Maybe<Scalars['String']>;
  /** The file extension from the file name. */
  extension: Maybe<Scalars['String']>;
  /** Indicates if this file is an image. */
  isImage: Scalars['Boolean'];
  /** Whether the file's contents were truncated. */
  isTruncated: Scalars['Boolean'];
  /** The programming language this file is written in. */
  language: Maybe<Language>;
  /** The gist file name. */
  name: Maybe<Scalars['String']>;
  /** The gist file size in bytes. */
  size: Maybe<Scalars['Int']>;
  /** UTF8 text data or null if the file is binary */
  text: Maybe<Scalars['String']>;
};
// ...
export type IssueConnection = {
  /** A list of edges. */
  edges: Maybe<Array<Maybe<IssueEdge>>>;
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<Issue>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type Label = Node & {
  /** Identifies the label color. */
  color: Scalars['String'];
  /** Identifies the date and time when the label was created. */
  createdAt: Maybe<Scalars['DateTime']>;
  /** A brief description of this label. */
  description: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  /** Indicates whether or not this is a default label. */
  isDefault: Scalars['Boolean'];
  /** A list of issues associated with this label. */
  issues: IssueConnection;
  /** Identifies the label name. */
  name: Scalars['String'];
  /** A list of pull requests associated with this label. */
  pullRequests: PullRequestConnection;
  /** The repository associated with this label. */
  repository: Repository;
  /** The HTTP path for this label. */
  resourcePath: Scalars['URI'];
  /** Identifies the date and time when the label was last updated. */
  updatedAt: Maybe<Scalars['DateTime']>;
  /** The HTTP URL for this label. */
  url: Scalars['URI'];
};
// ...
export type Language = Node & {
  /** The color defined for the current language. */
  color: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  /** The name of the current language. */
  name: Scalars['String'];
};
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars['ID'];
};
// ...
export type PageInfo = {
  /** When paginating forwards, the cursor to continue. */
  endCursor: Maybe<Scalars['String']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean'];
  /** When paginating backwards, the cursor to continue. */
  startCursor: Maybe<Scalars['String']>;
};
// ...
export type PullRequestConnection = {
  /** A list of edges. */
  edges: Maybe<Array<Maybe<PullRequestEdge>>>;
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<PullRequest>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type RepositoryOwner = {
  /** A URL pointing to the owner's public avatar. */
  avatarUrl: Scalars['URI'];
  id: Scalars['ID'];
  /** The username used to login. */
  login: Scalars['String'];
  /** A list of repositories that the user owns. */
  repositories: RepositoryConnection;
  /** Find Repository. */
  repository: Maybe<Repository>;
  /** The HTTP URL for the owner. */
  resourcePath: Scalars['URI'];
  /** The HTTP URL for the owner. */
  url: Scalars['URI'];
};
// ...
export type StargazerConnection = {
  /** A list of edges. */
  edges: Maybe<Array<Maybe<StargazerEdge>>>;
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int'];
};

/** Represents a user that's starred a repository. */
export type StargazerEdge = {
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
  node: User;
  /** Identifies when the item was starred. */
  starredAt: Scalars['DateTime'];
};

/** Things that can be starred. */
export type Starrable = {
  id: Scalars['ID'];
  /** Returns a count of how many stargazers there are on this object */
  stargazerCount: Scalars['Int'];
  /** A list of users who have starred this starrable. */
  stargazers: StargazerConnection;
  /** Returns a boolean indicating whether the viewing user has starred this starrable. */
  viewerHasStarred: Scalars['Boolean'];
};
// ...
export type UniformResourceLocatable = {
  /** The HTML path to this resource. */
  resourcePath: Scalars['URI'];
  /** The URL to this resource. */
  url: Scalars['URI'];
};
// ...
export type UserEdge = {
  /** A cursor for use in pagination. */
  cursor: Scalars['String'];
  /** The item at the end of the edge. */
  node: Maybe<User>;
};

/** Email attributes from External Identity */
export type UserEmailMetadata = {
  /** Boolean to identify primary emails */
  primary: Maybe<Scalars['Boolean']>;
  /** Type of email */
  type: Maybe<Scalars['String']>;
  /** Email id */
  value: Scalars['String'];
};

/** The user's description of what they're currently doing. */
export type UserStatus = Node & {
  /** Identifies the date and time when the object was created. */
  createdAt: Scalars['DateTime'];
  /** An emoji summarizing the user's status. */
  emoji: Maybe<Scalars['String']>;
  /** The status emoji as HTML. */
  emojiHTML: Maybe<Scalars['HTML']>;
  /** If set, the status will not be shown after this date. */
  expiresAt: Maybe<Scalars['DateTime']>;
  /** ID of the object. */
  id: Scalars['ID'];
  /** Whether this status indicates the user is not fully available on GitHub. */
  indicatesLimitedAvailability: Scalars['Boolean'];
  /** A brief message describing what the user is doing. */
  message: Maybe<Scalars['String']>;
  /** The organization whose members can see this status. If null, this status is publicly visible. */
  organization: Maybe<Organization>;
  /** Identifies the date and time when the object was last updated. */
  updatedAt: Scalars['DateTime'];
  /** The user who has this status. */
  user: User;
};
// ...
export type FeaturedUserQuery = { user: Maybe<{ contributionsCollection: { pullRequestContributions: { edges: Maybe<Array<Maybe<{ node: Maybe<{ pullRequest: (
              Pick<PullRequest, 'title' | 'url' | 'bodyHTML' | 'number' | 'state'>
              & { labels: Maybe<{ nodes: Maybe<Array<Maybe<Pick<Label, 'name' | 'color' | 'description'>>>> }>, repository: Pick<Repository, 'description' | 'nameWithOwner' | 'url' | 'stargazerCount' | 'forkCount'> }
            ) }> }>>> } }, repositories: { edges: Maybe<Array<Maybe<{ node: Maybe<(
          Pick<Repository, 'name' | 'url' | 'description' | 'stargazerCount' | 'forkCount'>
          & { languages: Maybe<{ edges: Maybe<Array<Maybe<{ node: Pick<Language, 'name' | 'color'> }>>> }> }
        )> }>>> }, pinnedItems: { edges: Maybe<Array<Maybe<{ node: Maybe<Pick<Gist, 'name' | 'url' | 'description' | 'stargazerCount'> | (
          Pick<Repository, 'name' | 'url' | 'description' | 'stargazerCount' | 'forkCount'>
          & { languages: Maybe<{ edges: Maybe<Array<Maybe<{ node: Pick<Language, 'name' | 'color'> }>>> }> }
        )> }>>> } }> };
// ...
export type RecentFollowingQuery = { user: Maybe<{ following: (
      Pick<FollowingConnection, 'totalCount'>
      & { nodes: Maybe<Array<Maybe<(
        Pick<User, 'login' | 'id' | 'bioHTML' | 'company' | 'location' | 'name' | 'websiteUrl' | 'twitterUsername'>
        & { avatar_url: User['avatarUrl'] }
        & { followers: Pick<FollowerConnection, 'totalCount'>, status: Maybe<Pick<UserStatus, 'emojiHTML' | 'message'>> }
      )>>> }
    ) }> };
// ...
export type SuggestedUsersQuery = { sindresorhus: Maybe<(
    Pick<User, 'login' | 'id' | 'bioHTML' | 'company' | 'location' | 'name' | 'websiteUrl' | 'twitterUsername'>
    & { avatar_url: User['avatarUrl'] }
    & { followers: Pick<FollowerConnection, 'totalCount'>, status: Maybe<Pick<UserStatus, 'emojiHTML' | 'message'>> }
  )>, emmabostian: Maybe<(
    Pick<User, 'login' | 'id' | 'bioHTML' | 'company' | 'location' | 'name' | 'websiteUrl' | 'twitterUsername'>
    & { avatar_url: User['avatarUrl'] }
    & { followers: Pick<FollowerConnection, 'totalCount'>, status: Maybe<Pick<UserStatus, 'emojiHTML' | 'message'>> }
  )>, rauchg: Maybe<(
    Pick<User, 'login' | 'id' | 'bioHTML' | 'company' | 'location' | 'name' | 'websiteUrl' | 'twitterUsername'>
    & { avatar_url: User['avatarUrl'] }
    & { followers: Pick<FollowerConnection, 'totalCount'>, status: Maybe<Pick<UserStatus, 'emojiHTML' | 'message'>> }
  )> };